FEAT!: Modularized the project in order to increase readability and OOP.
